+++ lib/bolt/target.rb
@@ -1,8 +1,154 @@
 # frozen_string_literal: true
 
 require 'bolt/error'
+require 'bolt/util'
 
 module Bolt
+  class Target2
+    attr_accessor :inventory
+
+    # Target.new from a plan initialized with a hash
+    def self.from_asserted_hash(hash)
+      inventory = Puppet.lookup(:bolt_inventory)
+      inventory.create_target_from_plan(hash)
+    end
+
+    # Target.new from a plan with just a uri
+    # rubocop:disable UnusedMethodArgument
+    def self.from_asserted_args(uri = nil,
+                                name = nil,
+                                target_alias = nil,
+                                config = nil,
+                                facts = nil,
+                                vars = nil,
+                                features = nil,
+                                plugin_hooks = nil)
+      inventory = Puppet.lookup(:bolt_inventory)
+      inventory.create_target_from_plan('uri' => uri)
+    end
+
+    # URI can be passes as nil
+    def initialize(uri = nil,
+                   name = nil,
+                   target_alias = nil,
+                   config = nil,
+                   facts = nil,
+                   vars = nil,
+                   features = nil,
+                   plugin_hooks = nil)
+      @name = name
+    end
+    # rubocop:enable UnusedMethodArgument
+
+    # Used for munging target + group data
+    def target_data_hash
+      {
+        'config' => @inventory.targets[@name]['config'],
+        'vars' => @inventory.targets[@name]['vars'],
+        'facts' => @inventory.targets[@name]['facts'],
+        'features' => @inventory.targets[@name]['features'].to_a,
+        'plugin_hooks' => @inventory.targets[@name]['plugin_hooks'],
+        'name' => @inventory.targets[@name]['name'],
+        'uri' => @inventory.targets[@name]['uri'],
+        'alias' => @inventory.targets[@name]['target_alias']
+      }
+    end
+
+    # features returns an array to be compatible with plans
+    def features
+      @inventory.features(self).to_a
+    end
+
+    # Use feature_set internally to access set
+    def feature_set
+      @inventory.features(self)
+    end
+
+    def vars
+      @inventory.vars(self)
+    end
+
+    def facts
+      @inventory.facts(self)
+    end
+
+    def to_s
+      safe_name
+    end
+
+    def config
+      @inventory.target_config(self)
+    end
+
+    def safe_name
+      @inventory.targets[@name]['safe_name']
+    end
+
+    def target_alias
+      @inventory.targets[@name]['target_alias']
+    end
+
+    def to_h
+      options.merge(
+        'name' => name,
+        'uri' => uri,
+        'protocol' => protocol,
+        'user' => user,
+        'password' => password,
+        'host' => host,
+        'port' => port
+      )
+    end
+
+    def host
+      @inventory.targets[@name]['uri_obj']&.hostname || @inventory.targets[@name]['host']
+    end
+
+    attr_reader :name
+
+    def uri
+      @inventory.targets[@name]['uri']
+    end
+
+    def remote?
+      @inventory.targets[@name]['uri_obj']&.scheme == 'remote' || @inventory.targets[@name]['protocol'] == 'remote'
+    end
+
+    def port
+      @inventory.targets[@name]['uri_obj']&.port || @inventory.targets[@name]['port']
+    end
+
+    # transport is separate from protocol for remote targets.
+    def transport
+      remote? ? 'remote' : protocol
+    end
+
+    def protocol
+      @inventory.targets[@name]['uri_obj']&.scheme || @inventory.targets[@name]['protocol']
+    end
+
+    def user
+      unencode(@inventory.targets[@name]['uri_obj']&.user) || @inventory.targets[@name]['user']
+    end
+
+    def password
+      unencode(@inventory.targets[@name]['uri_obj']&.password) || @inventory.targets[@name]['password']
+    end
+
+    def options
+      @inventory.targets[@name]['options']
+    end
+
+    def plugin_hooks
+      @inventory.targets[@name]['cached_state']['plugin_hooks']
+    end
+
+    def unencode(component)
+      Addressable::URI.unencode_component(component)
+    end
+    private :unencode
+  end
+
   class Target
     attr_reader :options
     # CODEREVIEW: this feels wrong. The altertative is threading inventory through the
@@ -87,6 +233,7 @@ module Bolt
         Set.new
       end
     end
+    alias feature_set features
 
     def plugin_hooks
       if @inventory