+++ spec/fixtures/modules/resource_types/lib/puppet/type/my_type.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+Puppet::Type.newtype(:my_type) do
+  @doc = "Empty type"
+end