+++ spec/pal/target_spec.rb
@@ -5,6 +5,7 @@ require 'bolt_spec/files'
 require 'bolt_spec/pal'
 
 require 'bolt/pal'
+require 'bolt/inventory'
 
 describe 'Target DataType' do
   include BoltSpec::Files
@@ -13,13 +14,13 @@ describe 'Target DataType' do
   before(:all) { Bolt::PAL.load_puppet }
   after(:each) { Puppet.settings.send(:clear_everything_for_tests) }
 
-  let(:pal) { Bolt::PAL.new(modulepath, nil) }
+  let(:pal) { Bolt::PAL.new(modulepath, nil, nil) }
 
   let(:target_code) { "$target = Target('pcp://user1:pass1@example.com:33')\n" }
 
   def target(attr)
     code = target_code + attr
-    peval(code, pal)
+    peval(code, pal, nil, Bolt::Inventory.new({}))
   end
 
   it 'should expose uri' do