+++ spec/pal/vars_spec.rb
@@ -20,7 +20,7 @@ describe 'Vars function' do
     }
   }
   let(:inventory) { Bolt::Inventory.new(data) }
-  let(:pal) { Bolt::PAL.new(modulepath, nil) }
+  let(:pal) { Bolt::PAL.new(modulepath, nil, nil) }
 
   let(:analytics) { Bolt::Analytics::NoopClient.new }
   let(:executor) { Bolt::Executor.new(1, analytics) }